refactor: renaming files to index and object destructuring
