+++ veascan-web/src/components/SnapshotAccordion/AccordionBody/index.tsx
