+++ veascan-web/src/components/SnapshotAccordion/AccordionTitle/index.tsx
